Do not set shared_preload_library property when it would be empty
When modifying a parameter group for a postgres database, we set the
`shared_preload_libraries` parameter when an extension that requires one is
requested. If none were required, we passed the empty string.

However, the RDS API will not accept an empty value for this parameter,
returning the below message. To avoid this, we now don't set the parameter if
no preload libraries are required.

    InvalidParameterValue: Empty value found in for parameter
    shared_preload_libraries

+++ rdsbroker/parameter_groups_source.go
@@ -78,8 +78,11 @@ func (pgs *ParameterGroupSource) setParameterGroupProperties(name string, servic
 
 	if aws.StringValue(servicePlan.RDSProperties.Engine) == "postgres" {
 		preloadLibs := filterExtensionsNeedingPreloads(servicePlan, provisionParameters.Extensions)
-		libsCSV := strings.Join(preloadLibs, ",")
-		dbParams = append(dbParams, rdsParameter("shared_preload_libraries", libsCSV, "pending-reboot"))
+
+		if len(preloadLibs) > 0 {
+			libsCSV := strings.Join(preloadLibs, ",")
+			dbParams = append(dbParams, rdsParameter("shared_preload_libraries", libsCSV, "pending-reboot"))
+		}
 	}
 
 	pgs.logger.Debug("modifying a parameter group", lager.Data{
